Move internal `base-radio`
