+++ packages/radix-ui-themes/src/components/radio.css
@@ -1,1 +1,1 @@
-@import './base-radio.css';
+@import './_internal/base-radio.css';
